Move parseToAst to its own module
