+++ lib/extractors/exported.js
@@ -2,23 +2,24 @@ var traverse = require('babel-traverse').default,
   isJSDocComment = require('../../lib/is_jsdoc_comment'),
   t = require('babel-types'),
   nodePath = require('path'),
-  fs = require('fs');
+  fs = require('fs'),
+  parseToAst = require('../parsers/parse_to_ast');
 
 /**
  * Iterate through the abstract syntax tree, finding ES6-style exports,
  * and inserting blank comments into documentation.js's processing stream.
  * Through inference steps, these comments gain more information and are automatically
  * documented as well as we can.
- * @param {Function} parseToAst funtiont that parses to an ast
  * @param {Object} ast the babel-parsed syntax tree
  * @param {Object} data the name of the file
  * @param {Function} addComment a method that creates a new comment if necessary
  * @returns {Array<Object>} comments
  * @private
  */
-function walkExported(parseToAst, ast, data, addComment) {
+function walkExported(ast, data, addComment) {
   var newResults = [];
   var filename = data.file;
+  var dataCache = Object.create(null);
 
   function addBlankComment(data, path, node) {
     return addComment(data, '', node.loc, path, node.loc, true);
@@ -75,7 +76,7 @@ function walkExported(parseToAst, ast, data, addComment) {
 
           var bindingPath;
           if (source) {
-            var tmp = findExportDeclaration(parseToAst, local, exportKind, filename, source.value);
+            var tmp = findExportDeclaration(dataCache, local, exportKind, filename, source.value);
             bindingPath = tmp.ast;
             specData = tmp.data;
           } else if (exportKind === 'value') {
@@ -128,9 +129,7 @@ function traverseExportedSubtree(path, data, addComments, overrideName) {
   }
 }
 
-var dataCache = Object.create(null);
-
-function getCachedData(parseToAst, path) {
+function getCachedData(dataCache, path) {
   var value = dataCache[path];
   if (!value) {
     var input = fs.readFileSync(path, 'utf-8');
@@ -148,9 +147,9 @@ function getCachedData(parseToAst, path) {
 }
 
 // Loads a module and finds the exported declaration.
-function findExportDeclaration(parseToAst, name, exportKind, referrer, filename) {
+function findExportDeclaration(dataCache, name, exportKind, referrer, filename) {
   var depPath = nodePath.resolve(nodePath.dirname(referrer), filename);
-  var tmp = getCachedData(parseToAst, depPath);
+  var tmp = getCachedData(dataCache, depPath);
   var ast = tmp.ast;
   var data = tmp.data;
 
@@ -202,7 +201,7 @@ function findExportDeclaration(parseToAst, name, exportKind, referrer, filename)
           if (exported === name) {
             if (source) {
               // export {local as exported} from './file.js';
-              var tmp = findExportDeclaration(parseToAst, local, exportKind, depPath, source.value);
+              var tmp = findExportDeclaration(dataCache, local, exportKind, depPath, source.value);
               rv = tmp.ast;
               data = tmp.data;
               if (!rv) {
