+++ lib/parsers/javascript.js
@@ -1,10 +1,10 @@
 'use strict';
 
-var babylon = require('babylon'),
-  _ = require('lodash'),
+var _ = require('lodash'),
   parse = require('../../lib/parse'),
   walkComments = require('../extractors/comments'),
-  walkExported = require('../extractors/exported');
+  walkExported = require('../extractors/exported'),
+  parseToAst = require('./parse_to_ast');
 
 /**
  * Left-pad a string so that it can be sorted lexicographically. We sort
@@ -20,31 +20,6 @@ function leftPad(str, width) {
     str = '0' + str;
   }
   return str;
-}
-
-function parseToAst(source) {
-  var opts = {
-    allowImportExportEverywhere: true,
-    sourceType: 'module',
-    plugins: [
-      'jsx',
-      'flow',
-      'asyncFunctions',
-      'classConstructorCall',
-      'doExpressions',
-      'trailingFunctionCommas',
-      'objectRestSpread',
-      'decorators',
-      'classProperties',
-      'exportExtensions',
-      'exponentiationOperator',
-      'asyncGenerators',
-      'functionBind',
-      'functionSent'
-    ]
-  };
-
-  return babylon.parse(source, opts);
 }
 
 /**
@@ -63,7 +38,7 @@ function parseJavaScript(data, options) {
   var addComment = _addComment.bind(null, visited);
 
   return _.flatMap(options.documentExported ? [
-    walkExported.bind(null, parseToAst)
+    walkExported
   ] : [
     walkComments.bind(null, 'leadingComments', true),
     walkComments.bind(null, 'innerComments', false),
@@ -73,12 +48,10 @@ function parseJavaScript(data, options) {
   }).filter(Boolean);
 }
 
-
-
 function _addComment(visited, data, commentValue, commentLoc, path, nodeLoc, includeContext) {
   // Avoid visiting the same comment twice as a leading
   // and trailing node
-  var key = JSON.stringify(commentLoc);
+  var key = data.file + ':' + commentLoc.start.line + ':' + commentLoc.start.column;
   if (!visited[key]) {
     visited[key] = true;
 
@@ -106,4 +79,3 @@ function _addComment(visited, data, commentValue, commentLoc, path, nodeLoc, inc
 }
 
 module.exports = parseJavaScript;
-module.exports.parseToAst = parseToAst;
