+++ lib/parsers/parse_to_ast.js
@@ -1,0 +1,30 @@
+'use strict';
+
+var babylon = require('babylon');
+
+var opts = {
+  allowImportExportEverywhere: true,
+  sourceType: 'module',
+  plugins: [
+    'jsx',
+    'flow',
+    'asyncFunctions',
+    'classConstructorCall',
+    'doExpressions',
+    'trailingFunctionCommas',
+    'objectRestSpread',
+    'decorators',
+    'classProperties',
+    'exportExtensions',
+    'exponentiationOperator',
+    'asyncGenerators',
+    'functionBind',
+    'functionSent'
+  ]
+};
+
+function parseToAst(source) {
+  return babylon.parse(source, opts);
+}
+
+module.exports = parseToAst;
